fix: make sure manifets files are always sorted
In the past there have been some cases that a new package was promoted
to the stage from downstream project and the manifest line was appended
at the end of the file via `actions/update-manifest` action.

This causes a conflict when Package Builder is executed manually in
Voulage where, although logically the content of the generated and
persisted manifests files are the same, the orders are not. And it
forces the package builder to continue attempting to build the packages.
Where in `testing` stage it shows its problem, because we are not able
to override existing published package.